(feat): add new methods to the services

+++ src/controllers/task.controller.ts
@@ -1,5 +1,6 @@
-import { Request, Response } from 'express';
-import * as TaskService from '../services/task.service';
+import { Request, Response } from "express";
+import * as TaskService from "../services/task.service";
+import { findUserById } from "../services/user.service";
 
 export const getTasks = async (req: Request, res: Response) => {
   const { userId } = req.params;
@@ -8,16 +9,47 @@ export const getTasks = async (req: Request, res: Response) => {
 };
 
 export const addTask = async (req: Request, res: Response) => {
-  const newTask = await TaskService.createTask(req.body);
+  const { userId } = req.body;
+  const exists = await findUserById(userId);
+  if (!exists) {
+    res.status(400).json({
+      errors: [{ field: "userId", message: "El usuario no existe" }],
+    });
+    return;
+  }
+
+  const taskData = {
+    ...req.body,
+    completed: req.body.completed ?? false,
+    createdAt: new Date().toISOString(),
+  };
+
+  const newTask = await TaskService.createTask(taskData);
   res.status(201).json(newTask);
 };
-
 export const editTask = async (req: Request, res: Response) => {
+  const taskId = req.params.id;
+
+  const exists = await TaskService.findTaskById(taskId);
+  if (!exists) {
+    return res.status(404).json({
+      errors: [{ field: "id", message: "La tarea no existe" }],
+    });
+  }
+
   await TaskService.updateTask(req.params.id, req.body);
   res.sendStatus(204);
 };
 
 export const removeTask = async (req: Request, res: Response) => {
+  const taskId = req.params.id;
+
+  const exists = await TaskService.findTaskById(taskId);
+  if (!exists) {
+    return res.status(404).json({
+      errors: [{ field: "id", message: "La tarea no existe" }],
+    });
+  }
   await TaskService.deleteTask(req.params.id);
   res.sendStatus(204);
 };
